+++ js/index.js
@@ -1,0 +1,3 @@
+'use strict';
+const native = require('node-gyp-build')(__dirname);
+module.exports = native;